fix: support non-RDMA local builds
