+++ python/Cargo.toml
@@ -21,8 +21,10 @@ name = "pegaflow-metaserver-py"
 path = "src/bin/metaserver.rs"
 
 [features]
-default = ["pegaflow-core/default", "pegaflow-server/default", "pegaflow-transfer/default"]
-cuda-13 = ["pegaflow-core/cuda-13", "pegaflow-server/cuda-13", "pegaflow-transfer/cuda-13"]
+default = ["cuda-12", "rdma"]
+cuda-12 = ["pegaflow-core/cuda-12", "pegaflow-server/cuda-12", "pegaflow-transfer?/cuda-12"]
+cuda-13 = ["pegaflow-core/cuda-13", "pegaflow-server/cuda-13", "pegaflow-transfer?/cuda-13"]
+rdma = ["dep:pegaflow-transfer", "pegaflow-core/rdma", "pegaflow-server/rdma"]
 
 [dependencies]
 pyo3 = { workspace = true, features = ["extension-module"] }
@@ -31,7 +33,7 @@ pegaflow-core = { workspace = true, default-features = false }
 pegaflow-proto = { path = "../pegaflow-proto" }
 pegaflow-server = { path = "../pegaflow-server", default-features = false }
 pegaflow-metaserver = { path = "../pegaflow-metaserver", default-features = false }
-pegaflow-transfer.workspace = true
+pegaflow-transfer = { workspace = true, optional = true }
 tokio.workspace = true
 tonic.workspace = true
 log.workspace = true
