+++ python/pegaflow/_bin_utils.py
@@ -1,6 +1,9 @@
 """Utility to locate Rust binaries bundled with pegaflow."""
 
+import os
 import shutil
+import site
+import sysconfig
 from pathlib import Path
 
 # pegaflow/ directory (where this file lives)
@@ -39,3 +42,26 @@ def find_binary(name: str) -> str:
         return found
 
     return name
+
+
+def _prepend_env_paths(env: dict[str, str], key: str, paths: list[str]) -> None:
+    paths = [path for path in paths if path]
+    if not paths:
+        return
+    current = env.get(key)
+    if current:
+        paths.append(current)
+    env[key] = os.pathsep.join(paths)
+
+
+def binary_env() -> dict[str, str]:
+    """Return an environment that can load binaries linked to this Python."""
+    env = os.environ.copy()
+    libdir = sysconfig.get_config_var("LIBDIR")
+    _prepend_env_paths(env, "LD_LIBRARY_PATH", [libdir] if libdir else [])
+    _prepend_env_paths(
+        env,
+        "PYTHONPATH",
+        [str(_MODULE_DIR.parent), *site.getsitepackages()],
+    )
+    return env
